refactor: use milan context to avoid repeated code
