--- src/styles/SignUpModal.css
@@ -1,41 +1,0 @@
-.signUpModalHeader > h1 {
-  font-family: Poppins, sans-serif;
-  color: #4e4e4e;
-  font-size: 1.5rem;
-  font-weight: 800;
-  text-align: center;
-}
-
-.button-wrapper {
-  display: flex;
-  flex-direction: column;
-  max-width: fit-content;
-  margin: 0 auto;
-}
-
-.modal-btn {
-  font-weight: 600;
-  color: #000000;
-  cursor: pointer;
-  margin: 20px;
-  text-align: center;
-  background-color: #e26959;
-  box-shadow: 0 0 50px -10px #9747ff14;
-}
-
-.btn:hover {
-  background-color: #e26959;
-}
-
-@media screen and (max-width: 525px) {
-  .signUpModalHeader {
-    justify-content: center;
-    align-items: center;
-  }
-
-  .signUpModalHeader > h1 {
-    font-size: 1.2rem;
-    margin-left: 1.4rem;
-    flex: auto;
-  }
-}
